added admin logic
